pyproject.toml
[tool.poetry]
name = "qc-notebooklm"
version = "0.1.0"
description = ""
authors = ["qiaochao <qiaochao@icourt.cc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.46.1"
flask = "^3.1.0"
sentence-transformers = "^3.2.1"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.16"
langchain-huggingface = "^0.0.3"
swarm = {git = "https://github.com/openai/swarm.git"}
aiofiles = "^24.1.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
srt = "^3.5.3"
sentencepiece = "^0.2.0"
datasets = "^3.1.0"
soundfile = "^0.12.1"
torch = "^2.5.1"
scipy = "^1.14.1"
openai = "^1.37.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pillow = "^11.0.0"
bark = "^0.1.5"
ipython = "^8.30.0"
suno-bark = {git = "https://github.com/suno-ai/bark.git"}
asyncio = "^3.4.3"
faster-whisper = "^1.1.0"
dashscope = "^1.20.14"
parler-tts = {git = "https://github.com/huggingface/parler-tts.git"}
pypdf = "^5.1.0"






[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
